Create Min Cost Climbing Stairs.py
I was not able to solve this problem, so I got help from Neetcode. So the basis of the problem is basically divide and conquer style. You first see the max value it takes from a index to go to the end, you add the minimum it costs you to the val in index itself. Then the problem gets simpler and simpler. We can apply the logic that
In this array [10, 15, 20, 5, 5, 0] -> 0 is place holder

we start at val 5 and see if 1 or 2 jump costs us more money. then we replace the value of the total cost of going from 5 to the desired place to index 3
This way we have total cost from index 3 to the end

now when we do the same thing for index 2 we see that any value works, and the new value would be 25

but at index 1 we see that moving 1 jump will force us to pay 30 to get to the end of the list while moving 2 places up gets us to the end of the list in 20.